Rename util file from dom to dom-sort
